프로그래머스 두 정수 사이의 합 (https://school.programmers.co.kr/learn/courses/30/lessons/12912)

+++ src/test/java/Lv1/programmers/Q12912/Q12912.java
@@ -21,11 +21,11 @@ public class Q12912 {
         );
     }
 
-    public static List<Integer> output() {
+    public static List<Long> output() {
         return asList(
-            12,
-            3,
-            12
+            12L,
+            3L,
+            12L
         );
     }
 }
